test: move out of unit
