+++ test/index.test.js
@@ -1,1 +1,9 @@
-import './unit/index.test.js'
+import './constructor.test.js'
+import './record-delta.test.js'
+import './histogram.test.js'
+import './clamped-averages.test.js'
+import './to-json.test.js'
+import './record.test.js'
+import './reset.test.js'
+import './plot.test.js'
+import './tick.test.js'
