--- test/unit/index.test.js
@@ -1,10 +1,0 @@
-import './constructor.test.js'
-import './histogram_ms.test.js'
-import './record-delta.test.js'
-import './histogram.test.js'
-import './clamped-averages.test.js'
-import './to-json.test.js'
-import './record.test.js'
-import './reset.test.js'
-import './plot.test.js'
-import './tick.test.js'
